eigendecomp float64

+++ kernels.py
@@ -10,8 +10,8 @@ def eigendecomp(K, y):
     n = K.shape[0]
     assert K.shape == (n, n)
     assert y.shape[0] == n and y.ndim == 2
-    K = torch.from_numpy(K).cuda()
-    y = torch.from_numpy(y).cuda()
+    K = torch.from_numpy(K).type(torch.float64).cuda()
+    y = torch.from_numpy(y).type(torch.float64).cuda()
     eigvals, eigvecs = torch.linalg.eigh(K)
     eigvals = eigvals.cpu().numpy()
 
@@ -19,7 +19,7 @@ def eigendecomp(K, y):
     eigvals /= n
     eigvecs *= np.sqrt(n)
 
-    eigcoeffs = (1/n) * eigvecs.T @ y.type(torch.float)
+    eigcoeffs = (1/n) * eigvecs.T @ y
     eigvecs = eigvecs.cpu().numpy()
     eigcoeffs = eigcoeffs.cpu().numpy()
     torch.cuda.empty_cache()
